refactor(jit-cache): separate provider artifact validation
